Moved code to markets home component
